docker api enforces tls from docker 1.13 onwards (#80)
* docker api enforces tls from docker 1.13 onwards

For docker 1.13 onwards, not using `dc.NewTLSClient` will fail with malformed http response.

Also, is there an option in dockertest v3 to support docker-machine? There was an option to support docker-machine in v2 but I can't seem to find it in v3. Providing such an option will allow us to use `dc.NewClientFromEnv()`.

* `NewTLSPool` for TLS endpoints

* NewClient vs NewTLSClient