fix: parse volume mounts using the same function as docker cli (#274)
volume mount paths should be of the form:
    `<src>:<dest>[:<options>]`

Docs:
https://docs.docker.com/storage/bind-mounts/#choose-the--v-or---mount-flag
https://docs.docker.com/storage/volumes/#choose-the--v-or---mount-flag

`src` can either be the name of a docker volume OR an absolute path on the host

MountParser assumes that it's parsing a local absolute
path for the `src` arg. However, as mentioned above, 
it's also legal to use a named docker volume. In addition, 
there can be the optional `options` section, with an
additional colon.

Rather than trying to capture all the edge cases coding
it ourselves, we just import the same function that
docker-cli uses to parse the volume.